requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.1.8
djangorestframework==3.12.4
drf-nested-routers==0.93.3
idna==2.10
pkg-resources==0.0.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
Pillow==7.0.0